Renamed event folder to event handler
