+++ manimlib/event_handler/event_dispatcher.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-from manimlib.event.event_type import EventType
-from manimlib.event.event_listner import EventListner
+from manimlib.event_handler.event_type import EventType
+from manimlib.event_handler.event_listner import EventListner
 
 
 class EventDispatcher(object):
